refactor: move command registry to top-level terok_shield.commands
The COMMANDS registry and its watch / simple_clearance runners lived
under cli/ — terok's integration layer imported the registry directly
from terok_shield.cli.registry, bypassing the adapter. Move all three
to top-level modules (commands.py, watch.py, simple_clearance.py); the
package __init__ already lazily re-exports COMMANDS / CommandDef /
ArgDef — repoint those at terok_shield.commands. cli/ keeps only the
argparse frontend.

Part of the cross-package command-registry normalization.
